bring over styles from .nav for .navbar-nav
since we're no longer using the .nav as a base class, we need to bring over some base styles for redoing browser list styles and setting flex in motion.

also brings with it some .nav-link styling. we're still using this global class, but with this small modification for alignment of content in responsive modes.